# API_BAJATEL/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Bajatel - Proyecto DAW",
        "description": "API REST para gestión de contratos de telecomunicaciones (Fibra, TV, Móvil)",
        "contact": {
            "email": "soporte@bajatel.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor de desarrollo"
        }
    ],
    "paths": {
...
        "/contratos/anadirServicioFibra/{id}": {
            "post": {
                "tags": [
                    "Servicios Contrato"
                ],
                "summary": "Añadir/actualizar servicio de fibra al contrato",
                "operationId": "49f3bf501a64464f3b16e52effb50723",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la opción de fibra",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Servicio añadido/actualizado"
                    },
                    "404": {
                        "description": "Usuario sin contrato"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/contratos/anadirServicioTV/{id}": {
            "post": {
                "tags": [
                    "Servicios Contrato"
                ],
                "summary": "Añadir/actualizar servicio de TV al contrato",
                "operationId": "3e390fa59b3eb916d9b37900cbed2c42",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la opción de TV",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Servicio añadido/actualizado"
                    },
                    "404": {
                        "description": "Usuario sin contrato"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/contratos/eliminarServicioFibra": {
            "delete": {
                "tags": [
                    "Servicios Contrato"
                ],
                "summary": "Eliminar servicio de fibra del contrato",
                "operationId": "de049a7ac3947f1dabd15096a9bb7909",
                "responses": {
                    "200": {
                        "description": "Servicio eliminado"
                    },
                    "404": {
                        "description": "Sin servicio de fibra"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/contratos/anadirLineaMovil/{id}": {
            "post": {
                "tags": [
                    "Servicios Contrato"
                ],
                "summary": "Añadir línea móvil al contrato",
                "operationId": "bc7a36a0e8a11c0a2441be81ff16f7dd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la opción móvil",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "numero_telefono": {
                                        "type": "string",
                                        "example": "612345678",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Línea añadida"
                    },
                    "404": {
                        "description": "Usuario sin contrato"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/contratos/eliminarLineaMovil/{numero}": {
            "delete": {
                "tags": [
                    "Servicios Contrato"
                ],
                "summary": "Eliminar línea móvil del contrato",
                "operationId": "75aae3f6e376c8ca5015df5764cc1ece",
                "parameters": [
                    {
                        "name": "numero",
                        "in": "path",
                        "description": "Número de teléfono",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Línea eliminada"
                    },
                    "404": {
                        "description": "Línea no encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum - Usar: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}